Option to record alternate_epoch but keep zero-based timestamps
Allow the option to set and record alternate_epoch but leave the
timestamps in all log files zero-based. This lets a user record the time
at which a job began with the desired clock_id while keeping nice
zero-based timestamps in the log files. (The zero-based timestamps are
easier to look at visually and also make the log files smaller, and
timestamps against the desired clock_id can be reconstructed by the user
if desired.)

The option log_alternate_eopch_clock_id is renamed to
alternate_epoch_clock_id. The new options relating to alternate_epoch
have not been in a released version of fio, so I expect this renaming
should be ok.

Signed-off-by: Nick Neumann nick@pcpartpicker.com