lib/rand: improve __fill_random_buf()
This won't be equivalent to what we have, but I _think_ the randomness
is good enough for this purpose.

While in there, fixup the coding style to conform to fio standards,
and simplify the setup and loop.

This improves performance by about 30% for me, tested on both aarch64
and x86-64.

Link: https://github.com/axboe/fio/pull/1417
Signed-off-by: Jens Axboe <axboe@kernel.dk>