engines/nfs: remove commit hook
The nfs ioengine does not need a commit hook because requests are
already committed when the queue hook is called. The queue hook calls
nfs_{pread/pwrite}_async and there is no separate commit step necessary
for the IO to be fully submitted.

This change also allows submission latencies to be reported because the
commit hook was not setting the io_u's issue_time.

Signed-off-by: Vincent Fu <vincentfu@gmail.com>