Fix logic determining if operating on single shared file
nrfiles=1 is not a valid decision point to determine whether
or not fio is operating on a single shared file. If numjobs>1
and filename is NOT specified, each job will create a single
non-shared file. add_file() is called differently when filename
is specified on the command line. If filename is specified,
then add_file() is called once from an options callback and
td->files_size==1. If filename is NOT specified, and each job
creates its own file, then td->files_size>=2 as filenames are
dynamically added. Change the logic to call setup_shared_file()
so that it is only called wheen td->files_size==1.

Signed-off-by: Chris Weber <weberc@netapp.com>