Fix IO might be hanging in batch completion storage system.
Some storage system could not complete the IOs until an entire batch
of IOs have been submitted. Therefore, the in-flight IOs will be hanging
in io_u_quiesce if they are less than a batch, and the next IO cannot be
scheduled. For this case, the option iodepth_batch_complete_omit for the
number of in-flight IOs that need not be retrieved, could be used to break
the hanging.

Signed-off-by: Fanghao Sha <shafanghao@gmail.com>