Change log_compression option from int to bool
The log_compression option no longer has an effect on when a chunk of
log entries is compressed in the background, but instead functions as a
boolean for whether or not to perform the compression. It was also used
to decide decompression chunk size after a job, but we can just use a
fixed chunk size as is done for the fio --inflate-log option.

As such, change the option from int to bool to avoid confusion. Note
that any existing usage of this option will need to be updated to
specify a bool instead of an int.

Signed-off-by: Nick Neumann <nick@pcpartpicker.com>