examples: fix bandwidth logs generation in disk-zone-profile.fio
Because the job name is set to "/dev/sdb", the bandwidth logs are
written to /dev/sdb_bw.1.log which is unexpected.

This commit changes the job name to "disk-zone-profile" so that the
bandwidth logs are written to "disk-zone-profile_bw.1.log".

Moreover, the `log_offset` option is enabled to log the I/O
offset. This is required to plot the HDD zones.

Signed-off-by: Christophe Vu-Brugier <christophe.vu-brugier@seagate.com>