Add log_compression_on option and deprecate log_compression
The log_compression option no longer has an effect on when a chunk of
log entries is compressed in the background, but instead functions as a
boolean for whether or not to perform the compression. It was also used
to decide decompression chunk size after a job, but we can just use a
fixed chunk size as is done for the fio --inflate-log option.

As such, deprecate the current option (to keep existing jobs working),
and add a new bool option named log_compression_on.

Signed-off-by: Nick Neumann <nick@pcpartpicker.com>